requirements.txt
streamlit
faiss-cpu
sentence-transformers
pdfminer.six
python-docx
python-pptx
pandas
transformers
torch